requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Retrieval
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
rank-bm25>=0.2.2

# Data processing
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Configuration & Utils
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0

# Monitoring (optional)
wandb>=0.15.0
mlflow>=2.8.0

# API clients (optional)
openai>=1.0.0
anthropic>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
